interview_bot/requirements.txt
# Core web framework
streamlit>=1.28.0

# AI and ML libraries
groq>=0.4.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# NLP libraries
nltk>=3.8.1

# Visualization
plotly>=5.15.0

# Environment management
python-dotenv>=1.0.0

# Database (SQLite is included in Python standard library)
# No additional database dependencies needed for SQLite

# Optional: For better performance and additional features
# torch>=2.0.0  # Uncomment if you want PyTorch backend for transformers
# transformers>=4.21.0  # Uncomment for additional transformer models
